Update Runtime & OrderSource API to Allow for multiple Images (#103)
* INT-5136 Create Shared App Specification

Going to have a shared app specification that all app implementations can use to define their apps.

* Add Basic Readme

* Update Dependencies

* Update yarn.lock

* Remove Shared Project

* Add things to ordersource & update runtime

These are the changes I was going to address in the shared code, but for now they can live in the ordersource api

* Too damn pretty.

* Update handler to take the request

For carriers and other implementation we want to allow them full access to the request body, headers, query params etc in order to format their request. For OrderSource this is a simple pass through of the body.

* Fix Image Routing

* Update sample to give example of multiple brandings

* Update Order Source & Carrier Runtime

Added the new images endpoint to both of them so we can unify how we get images in web api.

* Append transaction_id to error response

Based on feedback from Matt Burton, I think it would be a good idea to also include the transaction_id in the error response.

* Update Environment enum & how env values are set

* Wrap Sentry in if

Only run sentry if a DSN was provided.

* Fix pathing for util/storage

* Update docs

Matt Burton found an issue with docs pathing, this fixes it.

* Error handler should default send back error message

Without details provided the error handler should return the original error message in the details.

* Breakout files and separate internal classes/interfaces

In order to keep things cleanly separated I did some renaming, and added a new internal directory that includes classes & interfaces consumers should never need access to.