Add AppId To Manifest (#139)
* Add AppId To Manifest

It has been an issue to base the identity for a published application off of the package.json name property. So I added a new optional property appId to the manifest. This should hopefully reduce the number of times an app gets published as various different apps in our system.

* Fix linting errors

I am so glad we have a linter.

* Updates appId Automatically

Based on feedback from Josh, the appId will now be updated when you publish or do a connect info command.