bump dependencies to support redoc
We need webpack 5 and use the `resolve.fallback` option to handle
redoc's misuse of node builtins. Unfortunately this necessitates a bump
of next to version 11. Along with that comes some issues with the new
jsx transform in react, which I had to disable via babel.

Dependencies bumped

- next
- @mdx-js/react

Dependencies added for redoc

- buffer
- path-browserify
- process